Add jenkins, elk, chef, and internal DNS for each env (#54)
This makes a lot of changes:

- Moves chef so that we have a chef-server per environment, so that it is easy to move into another AWS account and not have cross-account or cross-tf dependencies.
- Move chef environments/roles into the kitchen dir so that the chef-server bootstrap can happen mostly automatically.
- Bring in the use of knife block to handle multiple chef servers.
- Adds in ELK stack
- Adds in jenkins as an deploy/ops automation platform, with jobs that will test terraform, deploy chef, and deploy code to the environment.
- Starts using berkshelf again
- Restructures tf so that we have different files for related systems (app, jenkins, elk), and main.tf is now just sort of global-ish stuff.
- Start using route53 for internal DNS in the VPC so that we can plug in CNAMEs for services (postgres, elk, Elasticsearch, etc) so that we can use the same names in all clusters rather than having to cut/paste hostnames into config.
- Restructure tf so that we don't have to hardcode subnets/VPC ids/etc
- Clean up env.sh so that it is easier to configure a cluster
- Update terraform outputs to emit URLs and environment variables that we can use for packer.
- Update packer so that it works in this new setup
- Update documentation to reflect all of the new processes

__Why__

This brings us significantly closer to a fully automated deployment/management platform.

We needed logging, so we got ELK
We needed a deployment/operations automation system, so we installed jenkins
Having a central chef server made it hard for us to delegate administration and provided a path for attackers to lateral into other environments.
Terraform was able to be cleaned up a bunch once we distributed shared systems, simplifying the config considerably.
We moved the manual chef provisioning step into terraform and brought back berkshelf, making this closer to being truly infrastructure as code.

__How__

Lots and lots of changes. Thanks to Justin, who managed to squash/rebase my megabranch into something that would merge in.