Add resource to generate htpasswd configuration files (#67)
__Why__

We were configuring this in multiple places. We created a file
(/opt/nginx/conf/htpasswd) with the contents from an encrypted data bag item. We
also have to set the user name and password for the basic auth in the sp-* apps
since they require those values to be set as ENV variables.

This change adds a chef resource to add a htpasswd file with configurable
filename, password, path, and username.

__How__

Create a chef resource that generates the htpasswd using the openssl command
line binary.