only allow users in the users databag to have sudo right (#83)
This also cleans up the way the users cookbook is used so that it works properly and only adds users who are in the proper environment group to the nodes.  I am not 100% sure that this actually removes sudo from ubuntu, but this code needs to get in for sure.