Readme Revamped ( Initial Draft) (#218)
**Why:**

There were some missing details that needed to be added
Read me needed to be re-organized to be able to provide clear instructions on how to build a VPC from beginning to end
There were some sections that needed to be removed and/or reorganized
**How:**
Use your favorite text editor. We don't judge!