Added scripts for bootstrapping a new environment (#252)
* Added environment with non user specific presets

Much of this configuration doesn't change user to user and is difficult to
find, so adding it should make this easier to get up and running.

* Remove use of whoami

My local machine does not have the same username as what I should have on the
infrastructure.

* Add GSA_FULLNAME and GSA_EMAIL

This is to avoid having to customize the environment script.

* Programatically set git branch that chef downloads

The chef server clones the identity-devops repo to get the chef cookbooks.
Assume that we want our current working branch to be deployed.

* Add TF_VAR_apps_enabled setting

* Added wrapper script to load environment

This does some basic checks to make sure you have the necessary environment
variables set.

* Remove all CIDR blocks besides GSA from whitelist

Even though we make this configurable, the only approved way to login is
from the GSA internal network, so that's the default.

* Fix rebase issue

There were updates to the way we get AWS power users which were not
reflected when we did the merge.

* Added better documentation

* Use spaces in load env

* Remove settings changes from load-env.sh

Because this file is sourced it messes with the settings of the calling
shell.

* Added scripts for bootstrapping a new environment

See https://github.com/18F/identity-private/issues/1810.

* Create zero elk users by default

This prevents an environment failing on initial bootstrap or having an
unexpected list of users because this file wasn't updated.   Eventually this
should be pulled out into the global config.

* Added kitchen/environments/sverch.json

* Make template_config_dbag.json valid JSON

* Minor style fixes

* Use spaces in check required files

* Minor error message fix

* Adding environment to test this pull request

See https://github.com/18F/identity-devops/pull/252

* Touch a file that terraform expects to exist

* Set first run placeholder for config databag key

* Change environment name

Our scripts fail if the environment name is too long

* Put random contents into chef user key

Apparently terraform complains if there's nothing there.

* Let the first terraform apply fail

* Turn off host key checking

This is not ideal, but we want the bootstrap script to finish

* Only create template files if they do not exist

So that multiple runs don't destroy an existing env

* Make deploy script support all terraform commands

The old script would fail if you passed other options because the
wrapper script didn't pass your arguments to terraform directly.  This
change ensures that terraform gets exactly the same arguments that you
pass to the script so that the wrapper can be used to take advantage of
terraform's full functionality.

* Update README

With this change, most of the steps in the README are automated, so call
that out up front.

* Add warning label to prevent run on existing env