Move power-users to explicit list in -private. (#344)
**Why?**

This eliminates an extra AWS API call at the beginning of every `deploy`
command that loads the environment. (This is annoying when you just want
to run `terraform show` or some other command that would otherwise be
fast.) It also makes it very explicit what the set of users ought to be.