Add --no-rebase to git auto update, ignore errors. (#362)
Other folks have encountered error messages from this git pull when
there are unstaged local changes that look like this:

    error: cannot pull with rebase: You have unstaged changes.
    error: please commit or stash them.

I haven't seen this myself (unstaged local changes work fine for me),
but I believe it happens when you set your git config to always rebase
instead of merging on git pull. You can mimic this by adding `--rebase`
to the command, so I expect that adding `--no-rebase` will make it go
away.

In addition to adding --no-rebase, make failure of git pull a warning
(in bold red) rather than a showstopper.

I'm worried that this will cause people to have outdated environments if
they are not able to pull identity-devops-private, but for now let's put
the onus on the user to discover if that's the case and fix it.