Add version info to environment deployment (#341)
* Add version info to environment deployment

Currently we don't have a good way of knowing which version of the terraform
configuration was used to deploy a given environment.  See
https://github.com/18F/identity-devops-private/issues/41 for more details about
why this is important.

See doc/process/releases.md for documentation on everything that this change
adds.

* Skip example cookbook in rake test

* Remove regression tests

These tests are for chef, so it doesn't make sense to run them when cutting
terraform versions.  When the chef repositories are all separate, this step can
be handled when those versions are bumped.

* Put regression test note in the right place

* Make environment version check a warning

* Make several small tweaks to version-info stuff.

- Move bundler/setup to be first require in Rakefile.
- Add trailing \n to VERSION.txt.
- Expand upon usage information for bin/get-version-info.sh.
- Move version info S3 path to be $PREFIX/version_info/$ENV.txt so
  they're all in a single directory.
- Fix some Ruby style nits.
- Add some example commands to the release process doc.
- Make terraform use of VERSION.txt independent of CWD.
- Add .terraform/modules/ to .gitignore.
- Make version_info.py executable.

* Force conversion of semver object to string

Because we are concatenating it with a string and writing it out to VERSION.txt.

* Make versioned deployment warning display always

And make it more visible