Match rspec chef-client version with real env (#379)
The version that rspec was downloaded didn't match the version that we have
deployed in our environments.

The upgrade of chef-client in our environments can now go in lock step with an
upgrade of all the cookbook chef-client versions, so we'll have regression
tests.  chef-client seems to have some differences between 12 and 13 that might
hit us when we do, so this will be important.