Add `lockout_period_in_minutes` setting (#388)
**Why**: We are no longer using rack-attack to rate limit how many
OTPs a user can request within a certain period of time. Instead, we
added logic in the main codebase to implement this. To keep things
simple, we replaced `otp_delivery_blocklist_bantime` with a
generic `lockout_period_in_minutes` that applies to both scenarios
that can cause a user to get locked out: requesting too many OTPs or
failing to enter the correct OTP more than 3 times in a row.