Make worker instances highly available. (#387)
* Make worker instances highly available.

Currently, the worker is a singleton instance. But it is in the critical
path of user requests because users cannot log in when it is down.

Create multiple worker instances (2 by default), while leaving the
existing DNS name and existing singleton instance untouched (except for
Name tag). This allows us to do a graceful migration without downtime.
Applying this change to an existing environment will just bring up an
extra worker instance.

* Add deprecation note on old worker.internal name.

* Use a round-robin list of AMIs.

* Fix worker node name to be unique.