Add a new script chef-databag. (#396)
* Add a new script chef-databag.

This script manages the process of downloading and uploading config
and user databags. It requires the user to have an ~/.ssh/config set up,
but otherwise assumes no dependencies on local knife setup, keys, etc.

It also provides a framework for more consistent user databags in the
future by synchronizing from S3.

* chef-databag: add download keys functionality.

Pull the functionality out of chef-create-user.sh that was used to
download the validator key and the data bag encryption key.

Be much more sophisticated about this behavior: if the keys already
exist but are unchanged, carry on happily. If they differ, prompt the
user for whether they want to keep a backup.

Also add -c /root/.chef/knife.rb to all the knife commands to ensure
that things will work even if we're running as a user like the ubuntu
user that our scripts will cause to have a broken knife.rb config.

Knife uses .chef/knife.rb from the CWD by default, which is a bad
default when you're running sudo to switch to a different user.

And be a little smarter with our backup prompts to give the user hints
about what is about to happen.

* chef-databag: Force correct .chef dir in upload

This was already used elsewhere, this was just a code path that wasn't updated.

This forces the knife upload command to use the correct .chef directory in /root
even if there is a local .chef directory in the directory that knife is run from
(which takes precedence even over a .chef directory in the user's home
directory).

* Add sync.sh back with a message

So that existing workflows will still see it