Create an auto scaling group for Jumphost provisioning. (#423)
* Add provision script to be run from user-data.

* Add chef download and install process.

* Have provision.sh install itself into PATH.

Also flesh out the skeleton for what chef-client --local will look like.

* Start adding a bootstrap TF module.

This module is used to generate user-data for ASG launch configurations
in a reusable way.

* Add an ASG for the jumphost.

Also tweak the way provision.sh is invoked to get rid of the wacky
self-installing behavior. If we ever want to rely on provision.sh being
installed in the AMI, this should change back to some other path.

Right now the ASG TF code is a bit of a hack due to a terraform bug
where it fails to process the create_before_destroy through the module
boundary. So if the launch configuration is inside the module but the
ASG is outside the module, it will attempt to delete the launch
configuration before updating the ASG, which fails.

This isn't fixed in latest TF master, so deduplicating this will have to
wait.

* Fix a bunch of bootstrapping bugs.

- Add an IAM role for the jumphost so that it can access the secrets
  bucket.
- Install git.
- Fix up a hack to make the git clone SSH actually use the right key.

* Drop timestamp from version-info.

* Add github to ssh_known_hosts.

The config is now at the point where the auto scaling group provisioner
can get through the full chef install and clone, and we are now ready to
run chef.

* Run provision against new chef-client.rb.

* Also install berkshelf and run it in provision.sh.

This avoids needing to vendor the cookbooks in the repo itself.

Also add a flag to configure the kitchen / chef repo subdirectory.

* Make jumphost ASG desired count a variable, def 0.

* Evidently berks needs HOME to be set.

* Fix up berks vendor.

* Document TODOs and split out cruft.

* Convert all of the bootstrap inputs to variables.

These are going to be repeated once for each role of server, so add them
to variables but add defaults.

Also add some (currently unused) variables that will be used when
provision.sh is eventually used to clone identity-devops in addition to
identity-devops-private.

* Latest berkshelf doesn't work with Chef 12.15.