Add helper to JSON.parse build_env as needed.
In config_loader, add a new helper `ConfigLoader.load_json`, which calls
`JSON.parse` on the secrets data if and only if it is loaded from
citadel. This helps us handle cases where we are currently relying on
nested hashes or other complex data in encrypted data bags.

Use this new helper to handle the jenkins build_env.