Add ConfigLoader.load_config_or_nil helper.
In a small number of cases we rely on the ability to have a Citadel downloaded
secret return nil so that we can use default attributes instead.

Add a helper `ConfigLoader.load_config_or_nil` that swallows HTTP 404 errors
(and prints a warning) to enable this functionality for this small number of
cases. Because the warning is easy to miss, don't swallow the error if
node.chef_environment is "prod": in that case always fail, since production
should never be using default secrets.