Jenkins: use -o to avoid changing chef run list. (#442)
* Jenkins: use -o to avoid changing chef run list.

In the jenkins IDP jobs, currently we mutate the chef run list when we
run chef. This can be confusing and leave nodes in an unexpected state
if the job fails partway through.

Instead use `chef-client -o` so that the jenkins runs deploy the IDP
code as expected and no run list changes are needed.

* Pass app name directly to new relic resource.

Instead of passing the app_friendly_name implicitly through a node.set,
just pass it as a property on the new relic config resource.