secret_key_base is not a thing that exists.
Pretty sure that this is broken/not working in master, because no
environment has an encrypted data bag key "secret_key_base".

It also appears that we're passing these values as environment variables
shipped through the nginx config, in addition to values passed through
secrets.yaml. It's a bit of a mess.