Run chef on identity-devops-private first.
Also add a long comment explaining the tradeoffs. Ideally we'd like to
ensure that we have no accidental dependencies in identity-devops on
identity-devops-private. But given that identity-devops bootstrapping
takes so long, it's important to create unix users early so that we can
SSH in to troubleshoot when things go sideways.

As a result of this change, we also no longer ignore the exit status
from the chef-client run on identity-devops. If bootstrapping fails,
/run/cloudinit-finished will not be created.