Move run list / role configs into chef-client.rb.
Previously we would have provision.sh specify the environment and run
list when invoking chef-client. But ideally we want chef-client.rb to be
fully self-sufficient so that we can invoke chef-client by hand with the
same result.

In the cloud-init boostrapping, write out a chef attributes JSON file
that we use to drive the run list. In the chef-client.rb, set the
environment and reference this attributes file.

There should be a corresponding change to the chef-client.rb in
identity-devops-private.

Also, instead of using the auto-scaled sentinel file in cookbooks to
detect whether we're running in an ASG, switch to using chef node
attributes that are set by the JSON attributes file.

In addition, prefix the launch configuration name with the git ref that
we're using to clone the main identity-devops repo so that you can see
from the launch configuration what git ref the instances were
bootstrapped from.