Fix jumphost tests by downgrading berkshelf. (#473)
For reasons that I do not yet understand, `berks update` using Berkshelf
5.x or 6.x does not correctly follow cookbook dependencies. This causes
needed cookbooks to be missing on the kitchen EC2 instance.

Downgrade to Berkshelf 4.x so that we generate a working Berksfile.lock.

Also add the Berksfile.lock to git. We can paper over some of these
dependency issues just by having a Berksfile.lock. (This is why we don't
yet have this dependency issue with the repo toplevel Berksfile, because
there is a committed Berksfile.lock from Berkshelf 4.x that contains the
correct dependencies.)

Also install kitchen-sync, which slightly speeds up the glacial pace of
the kitchen-ec2 cookbook upload process by using net-sftp rather than
net-scp, which is incredibly slow with a large number of files.