Add enable flag to all aws_instance resources (#470)
* Add enable flag to all aws_instance resources

This will allow both for creation of the continuous integration VPC and for easy
transition to auto scaling groups.

For the continuous integration VPC, all nodes will be disabled to create a VPC
with only a database and a cache.  This will allow test kitchen to spin up
instances in this VPC and test that they work, even if they assume certain
base components exist.

For the transition to autoscaling groups, this allows us to gradually transition
by only moving one node at a time, and by migrating per environment.  Since the
default is no change, the configuration must be explicitly changed to remove
these instances.

* Fix issues to get basic plan working

Needed to not check for required files if those instances didn't exist, and
catch some other trailing issues.

* Don't try to get a certificate if we have no alb

* Handle possibly unset TF_VAR variables.

* Use 1 and 0 instead of true or false

Terraform does not support booleans, only numbers and strings.  So if we try to
use true it's inconsistent as to whether it's treated as a true string or the
number 1.

See
https://blog.gruntwork.io/terraform-tips-tricks-loops-if-statements-and-gotchas-f739bbae55f9.

* Correctly treat flag as number in acme configs

* Update CI EC2 Test Kitchen config

__Why__

* we want integration tests to run in the CI VPC, not the dev VPC

__How__

* change the test-kitchen ec2 configuration to point to ci vpc instead of dev

* We only have 1 of the deprecated worker DNS name.