Make idp role run role[base].
As part of the move to auto scaling groups and test kitchen friendly
configuration, we are removing all of the custom and manual run lists
that are currently used to provision hosts.

Instead, each type of server ("jumphost", "idp", "worker", etc.) will
use role[TYPE] to bootstrap itself and pull in everything that is
needed.

Rename the existing "idp" role to "idp_only", since it deploys the IDP
rails app but doesn't actually bootstrap the machine. Add a new role
"idp" that runs role[base] and role[idp_only].