Chmod +X rather than +x on /srv/.
There is no reason to make all files in /srv/ executable if our goal is
to make directories readable. Use chmod +X, not +x so that we don't
accidentally make all files executable.

Also add a variety of TODO comments indicating that tossing around
`chown -R` everywhere is **not** a best practice.