ASG recycle: use grace as minimum spindown time.
ASGs have pretty questionable behavior: where if you scale in an ASG
while instances are still in the health check grace period, Amazon will
consider the still-booting instances healthy. So under most termination
policies, especially OldestInstance, this causes the ASG to terminate
old healthy instances before the new instances have actually passed the
ELB health check.

To reduce the risk of this, use the ASG's health check grace period as a
lower bound on the spin down delay that we wait for in asg-recycle.sh.