Run passenger restart hack as a handler.
For reasons that I couldn't figure out, we need to restart nginx at the
very end of the chef run, after everything else has happened. The
deferred `service passenger restart` was not doing the trick.

So create a handler that runs after the whole chef run has succeeded
that checks whether Passenger is running and restarts it/nginx if not.

This fixes the test added in b76845232864c73d1ab9a5793dc320482e4c9b2d !