Move test files into common dir and remove chefignore (#508)
* Move files into common dir and remove chefignore

There are some files that do not change, such as the Gemfile and the user_data
passed into cloud-init.  Both of these are simply part of the test boilerplate
and not really different per node.

Once our security group and IAM role names are more standardized based on the
role name, we should be able to move .kitchen.cloud.yml into this common
directory too.

The environment file used by chef should also not change, and should match what
we actually deploy nodes with in this environment, so this was also factored out
(but into the main kitchen rather than the common test directory).

The Berksfile should already be symlinked from the top level.

Chefignore does not serve a purpose here, as it is only used by knife to upload
cookbooks.

* Fix race between cloud-init and test kitchen (#510)

Creating the sandbox directory and mounting the volumes must happen before test
kitchen tries to use them, otherwise the test will fail.