Fix bad bug with ssh-instance + ssh sockets. (#526)
It uses the name_tag as the name passed to SSH as the hostname when a
jumphost is in use. This causes SSH to use the Name tag as the
identifier when creating a controlmaster socket if controlmaster is
enabled. This, in turn, will cause it to actually SSH to the same server
if a control socket exists with a server with the same Name tag (e.g.,
for autoscaled instances they are all called `asg-env-idp`).