Add support for testing the cloud-init configuration with test kitchen (#520)
* Template cloud-init files using ERB

This will allow us to reuse these configuration files in our test kitchen
integration tests.

* Add test kitchen boilerplate for cloud init

This isn't yet using the real bootstrap, but at least passes the user_data into
test kitchen.

* Split cloud init config into separate phases

The provisioning and the base system configuration are logically separate.  This
will allow us to reuse these in our integration tests.

* WIP Trying to move all kitchen.cloud.yml to ruby

* Finish making the kitchen cloud conf a template

* Remove extra header in test kitchen YAML

Now that this is all generated, this is unnecessary

* Fix type mismatch in compression level

* Comment out some things that aren't working yet

Things still need to be done:

- Figure out how to get test kitchen to wait for cloud-init to complete
- Don't install chef omnibus, let the provisioner do it

* Rename test kitchen boilerplate script

To more match what its purpose actually is.  Also leave a comment about the fact
that it is strange that we need this.

* Correctly pass in kitchen subdir berks location

It's different in identity-devops and identity-devops-private.

* Remove unnecessary assignment