Skip basic auth for healthchecks.
This will allow us to remove the weird 401/200 HTTP status code split in
the future. The ALB should always be able to reach the health check and
get a meaningful value.

We could implement this in a different way by trusting the
X-Forwarded-For header in some way, but this seems like a better
stopgap.