Fix notify behaviour in cert install resource (#560)
* Fix notify behaviour in cert install resource

Chef notify gets triggered if anything in the resource changes, but it treats
log messages as a change all the time.  Removed the log message since it is
redundant with what is logged by the file resource anyway.

* Add .inspect to debug output

This makes more sense than to_s for debugging, which is the default.