Add per account configuration support (#555)
* Add per account configuration support

We have some account wide configuration that is not per environment, but at the
moment it just reuses whatever environment file is passed in as an environment
name.  This could cause issues if two people for example run terraform-common
passing in two different things as the environment name.

* Bump env compat version for safety

* Address review comments

Do not set a default state bucket, and use cleaner way to get account id.