Change Name tag on app db to reflect purpose. (#436)
* Change Name tag on app db to reflect purpose.

This makes it less confusing to identify the app RDS instance. Currently
it is tagged "login-$ENV". With this change it will be "login-$ENV-app".

* Add outputs for the actual RDS DB endpoint hosts.

We can't use the postgres.login.gov.internal DNS names in idp/app config
because we verify SSL certificates, and the certificate name only
matches the actual RDS endpoint hostname.

These outputs provide a shortcut to seeing where to connect to the RDS
DB instances.