Fix Citadel config for prod. (#583)
ConfigLoader.load_config_or_nil deliberately raises an error in prod if
it can't find a Citadel config file. This is because in several places
we use load_config_or_nil in non-prod environments to fall back on
default values for secrets that are stored in this git repository, which
would be a catastrophic failure in prod.

Don't write out the ~ubuntu/.bash_profile at all in prod. We should not
be using the ubuntu user period.