Switch bootstrap to use acct-agnostic S3 bucket. (#612)
We currently hardcode the expected default bucket name, but under the
new schema these buckets will be suffixed by AWS account ID and region,
which is a potentially large set of bucket names.

Make passing the S3 URL mandatory for now. Once we upgrade to the
upcoming next Terraform 0.10.* release, we'll be able to use local
variables to compute this on the fly in a reasonable way.

This is a prerequisite for standing up servers in the nonprod AWS
account.