Move to a new CI/CD system (#616)
This repository has been using Travis CI, however access to Travis CI is now turning off. When you need to use this repository again, convert it to a CI/CD system which is in the ITSP such as Circle CI.

Sorry to do this in a PR, I don't see an option to create an issue.