Fix sp_certificate and sp_private_key strings.
These are currently strings that contain literal '\n' characters,
working under the assumption that the content will be splatted directly
into a YAML file later where it will be interpreted as a newline
character.

This is failure-prone. Instead treat the content as an actual ruby
string and thread it properly through YAML serialization without any
potential for syntax errors or incorrect escaping.

Ideally we shouldn't have even sample private keys in the repo at all,
but that's a task for another day.