ssh-instance: Interactively prompt to select host.
Add some functionality to interactively select an instance among several
options. This makes it easier if you just want to SSH to an instance but
you aren't sure or don't want to have to specify in advance which ones.

Make the interactive prompt the default if stdin/stderr are connected to
a TTY, otherwise continue requiring the name tag to be unique.

For example, you can now do this:

    $ bin/ssh-instance asg-staging-idp
    Found multiple instances named "asg-staging-idp": (Use arrow or number (1-4) keys, press Enter to select)
    ‣ 1. i-07c00bc873a71cee2, us-west-2a, 2018-01-04 18:13 UTC (149.0 hrs ago)
      2. i-0047492b51ad5b97b, us-west-2a, 2018-01-04 18:13 UTC (149.0 hrs ago)
      3. i-037767b689ad6497b, us-west-2b, 2018-01-04 18:13 UTC (149.0 hrs ago)
      4. i-050a9510402c8b40e, us-west-2b, 2018-01-04 18:13 UTC (149.0 hrs ago)

    <enter>

    I, INFO -- ssh-instance: SSH to "asg-staging-idp" (i-07c00bc873a71cee2)