Add diff-deploy for interactive TF plan + apply.
The diff-deploy script is a wrapper around the existing deploy script,
but it adds some nice features to show a pretty terraform diff and then
interactively prompt for approval to continue. If the user confirms,
then diff-deploy will apply the terraform plan exactly as generated,
which removes any possibility of accidental differences due to
intervening environment / variable / code changes.