Add a variable to control new vs old log bucket.
This allows us to gracefully move to the new log bucket naming scheme on
a per-environment basis. We stick with the old naming scheme by default,
but switch to the new one if TF_VAR_legacy_log_bucket_name is set to 1.

As a result of this change we can gracefully move environments from
login.gov to identitysandbox.gov without requiring code changes in
identity-devops.