Use a deterministic cache dir for openssl recipe.
The OpenSSL compilcation currently saves intermediate artifacts
(tarballs, compiled source tree) under the main Chef file cache path.

This is problematic because we are using the same recipe in the base AMI
and in identity-devops cookbooks, which need to share the cache dir in
order to avoid redoing all the compilation steps. But the file cache
path will differ depending on how chef is being run.

In the base image build, the Chef::Config['file_cache_path'] is:
    /tmp/packer-chef-solo/local-mode-cache/cache/

In the identity-devops chef runs, Chef::Config['file_cache_path'] is:
    /etc/login.gov/repos/identity-devops/kitchen/local-mode-cache/cache/

Instead of using either of these directories, now use `/var/cache/chef`
to store intermediate artifacts for OpenSSL compilation.