Pave the way for a future deploy/build script.
We would like to change the interface between identity-devops and
identity-idp so that the chef cookbook doesn't need to know details
about how the idp configures itself. Decoupling these allows us more
flexibility when rolling out new releases that change the way we do
app internals like compiling static assets or updating the database.

Ideally the idp repo should know how to build itself, and the devops
repo should only call the consistent interface that is shared by all of
our services.

Add hooks to call a `deploy/build` script that when it exists will be
the home for steps like `bundle install`, `yarn install`, and
`rake assets:precompile`. These hooks are conditional on whether that
script exists in order to allow for a graceful migration to the new
interface.