Revert npm => yarn move, rely on deploy/build.
We can't move to yarn on older versions of the idp, so just use the
existence of deploy/build to handle the cutover.

If deploy/build does not exist, we will run npm install as usual. If
deploy/build does exist, then run it. (It will in turn run yarn
install.)