Tweak invalid user alert message for accuracy.
The "ssh ports may be exposed!" message was alarmist and inaccurate.
This alert indicates only that an invalid user attempt was detected. If
we want an alert for whether connections happened from an unexpected IP
address range, we should add a new check that looks at the IP addresses
directly.