Enable static EIPs for jumphost/workers in prod.
This has been baking for quite some time in QA, so I think we're ready
to roll it out everywhere. Static IPs are useful whenever we need to
whitelist services using a static CIDR block.

Right now, we use static IPs on the jumphosts to firewall access to
Redshift. (This will not be necessary once we do VPC peering with the
Analytics VPC, which will allow granting access by security group ID
instead.)

We also plan to use static IPs on the worker servers to firewall access
to AAMVA.

The IP ranges for these hosts have already been configured in secrets
and are documented in the Google doc with IP address allocations.

Closes: https://github.com/18F/identity-devops-private/issues/666

See also: https://github.com/18F/identity-devops-private/issues/570