Add more helpful error when VPC doesn't exist.
This is usually because you're looking in the wrong AWS account.

Ideally the tooling would be smart enough to have a config mapping of
env name => AWS account profile name, so that the user wouldn't have to
provide this themselves every time. But for now this makes it a little
less confusing.