Fix up kitchen tests to be working again.
The idp/worker kitchen tests were broken by the deploy provisioning
changes that switched to invoking the deploy/build script in the
identity-idp repo.

This was because of some weird behavior where something is running
bundler as root with HOME=/home/ubuntu, which creates a
/home/ubuntu/.bundle/cache directory as root. This in turn causes the
deploy/build script's bundle install as the ubuntu user to fail because
it can't create anything in this directory.