Don't clobber identity-idp/bin directory.
Previously we would delete the entire identity-idp/bin directory and
symlink it to /srv/idp/shared/bin, where we installed the bundler
binstubs from running bundle install. This was extremely confusing if
you expected files from identity-idp/bin to be present, and I believe we
didn't use any of the binstubs anyway. (All the callers I could find use
bundle exec foo instead.)

This also conflicts with recent refactors of the deploy/activate script
which place a script in bin/activate that we need to call:

https://github.com/18F/identity-idp/pull/2069