User dn/uuid mapping
**Why**:
We want to know if a user is returning to the system.

**How**:
We record the DN from the public certificate and send along a UUID for
use elsewhere. We encrypt the DN so a dump of the database doesn't
make it easy to match up identities with records in identity-idp.