Endpoint to receive client cert from piv/cac
**Why**:
We want to accept the client certificate from the piv/cac
and redirect back to the IdP with a token that can give
the IdP information about the piv/cac without leaking
more information than required.

**How**:
We accept the client certificate from Nginx and validate
it against the authorities we have in-hand. If valid, we
return a token for success. Otherwise, we return a token
with an error.