Rake tasks to dump certs and graph relationships
**Why**:
We want to share the trusted set of certificates with
Nginx.

**How**:
A rake task to dump the list of verified certificates that
we can trace back to a trusted root. As part of this, we
also have a rake task to create a plantuml diagram of how
the certificates are related to the trusted root(s).