deploy: Bail out if env name is unknown.
Previously the deploy script would try to carry on anyway if it was
given an environment name for which there is no config in
identity-devops-private.

In practice we never want to run deploy with an unknown environment
name. In 99% of cases this is just user error or a sign that something
has gone horribly wrong.

In the 1% of cases where you are trying to clean up an existing borked
environment, you can create an empty environment file to work around
this change.