Add scale-in-old-instances to help asg recycle.
A major nuisance currently with recycling auto scaling groups is that it
isn't obvious when it's safe to remove scale-in protection. This script
automates some of the pain away by checking that the new instances pass
ASG health checks and that the health check grace period has elapsed.

It does *not* check ALB health checks or do anything else to determine
whether the instance is truly working.