Fix filtering for only ELB errors option.
Previously, the elb-find-errors script would find HTTP 5xx errors
returned by the target group even when --elb-errors was set, which is
supposed to filter them out. This is because the HTTP status code for
the ELB will match the target group when the target group returns an
error.

Now correctly filter only for log lines where the target group status
code is "-".