Set all env vars from /etc/environment in ENV.
Various commands executed during the chef run need the environment
variables that we're trying to set globally from /etc/environment. But
subprocesses of the chef run where we set /etc/environment for the first
time won't have them set by default.

Hack around this by setting each of those variables in ruby ENV so that
they are set for chef and all subprocesses that follow this cookbook.