Fix configure_state_bucket for ancient Bash 3.
Bash 3.2 has a bad bug where if a shell command run by `command` exits
false and the script is run with errexit, the entire script will
immediately exit even if the command is inside an if statement or has a
following `||` conditional.

Apple ships with /bin/bash running Bash 3.2, released 12 years ago. (!!)
So work around this by switching to `env`, which doesn't have this
issue, instead of the built-in `command`.