Assume chef is already installed in the AMI.
Previously we would download chef at runtime (twice) and reinstall it
from a debian package. This is no longer useful since
identity-base-image and identity-devops now use the same version of
chef. Instead, assume chef is already installed in the AMI.

This will make provisioning a little faster and reduce the duplication
here. If we want to roll out a new version of chef in the future we can
always reintroduce these variables.